+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ versión, aunque sí su incorporación en la rama principal de trabajo, generalm
 ### Versión 0.1.4 2022-10-25
 
 - Se hacen más estrictas las correcciones de los regímenes usando expresiones regulares.
+- Se agregan pruebas para el método `Regimen::setFechaAlta`.
 
 ### Versión 0.1.3 2022-10-25